+++ packages/rpmtw_api_client_flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- add global auth token
+- get user info token option
+
 ## 0.0.1
 
 - Add avatar widget to user profile